Move zipkin translator to pkg/translator/zipkin
