+++ pkg/translator/zipkin/go.mod
@@ -1,0 +1,17 @@
+module github.com/open-telemetry/opentelemetry-collector-contrib/pkg/translator/zipkin
+
+go 1.16
+
+require (
+	github.com/census-instrumentation/opencensus-proto v0.3.0
+	github.com/google/go-cmp v0.5.6
+	github.com/jaegertracing/jaeger v1.25.0
+	github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal v0.0.0-00010101000000-000000000000
+	github.com/openzipkin/zipkin-go v0.2.5
+	github.com/stretchr/testify v1.7.0
+	go.opentelemetry.io/collector v0.33.1-0.20210820002854-d3000232f8f6
+	go.opentelemetry.io/collector/model v0.33.1-0.20210820002854-d3000232f8f6
+	google.golang.org/protobuf v1.27.1
+)
+
+replace github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal => ../../../internal/coreinternal
