+++ pkg/translator/zipkin/internal/zipkin/attributes.go
