+++ pkg/translator/zipkin/zipkinv1/consumerdata.go
