+++ pkg/translator/zipkin/zipkinv1/testdata/zipkin_v1_error_batch.json
