+++ pkg/translator/zipkin/zipkinv1/testdata/zipkin_v1_local_component.json
