+++ pkg/translator/zipkin/zipkinv1/testdata/zipkin_v1_thrift_single_batch.json
