+++ pkg/translator/zipkin/zipkinv2/marshaler.go
