+++ pkg/translator/zipkin/zipkinv2/testdata/zipkin_v2_notimestamp.json
